WIP #286 plans
