+++ apstools/plans/__init__.py
@@ -1,2 +1,43 @@
 
+"""
+Plans that might be useful at the APS when using Bluesky
+
+.. autosummary::
+   
+   ~streams.addDeviceDataAsStream
+   ~command_file.execute_command_list
+   ~command_file.get_command_list
+   ~alignment.lineup
+   ~scans.nscan
+   ~command_file.parse_Excel_command_file
+   ~command_file.parse_text_command_file
+   ~command_file.register_command_handler
+   ~command_file.run_command_file
+   ~snapshots.snapshot
+   ~sscan.sscan_1D
+   ~command_file.summarize_command_file
+   ~alignment.TuneAxis
+   ~alignment.tune_axes
+
+"""
+
+#-----------------------------------------------------------------------------
+# :author:    Pete R. Jemian
+# :email:     jemian@anl.gov
+# :copyright: (c) 2017-2020, UChicago Argonne, LLC
+#
+# Distributed under the terms of the Creative Commons Attribution 4.0 International Public License.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+#-----------------------------------------------------------------------------
+
+# import logging
+# logger = logging.getLogger(__name__)
+
+from .alignment import *
+from .command_file import *
 from .motors import *
+from .scans import *
+from .snapshots import *
+from .sscan import *
+from .streams import *
